Remove unused endpoints
