+++ MovieApi/MovieApi/Controllers/MovieController.cs
@@ -1,7 +1,5 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using MovieApi.Models;
-using MovieApi.Models.GetMovieDetail;
-using MovieApi.Models.GetMovieList;
 using MovieApi.Services;
 
 namespace MovieApi.Controllers
@@ -19,6 +17,7 @@ namespace MovieApi.Controllers
             _movieService = movieService;
         }
 
+        
         [HttpGet("movies")]
         public async Task<ActionResult<GetMovieListResponse>> GetMovieListAsync()
         {
@@ -30,32 +29,6 @@ namespace MovieApi.Controllers
             }
 
             return Ok(movies);
-        }
-        [HttpGet("movies2")]
-        public async Task<ActionResult<GetMovieListResponse2>> GetMovieListAsync2()
-        {
-            var movies = await _movieService.GetMovieListAsync2();
-
-            if (movies == null)
-            {
-                return NotFound("No movies found.");
-            }
-
-            return Ok(movies);
-        }
-
-
-        [HttpPost("movie")]
-        public async Task<ActionResult<GetMovieDetailResponse>> GetMovieDetailAsync([FromBody] GetMovieDetailRequest request)
-        {
-            var movieDetail = await _movieService.GetMovieDetailAsync(request);
-
-            if (movieDetail == null)
-            {
-                return NotFound("No movie detail found.");
-            }
-
-            return Ok(movieDetail);
         }
     }
 }
