--- MovieApi/MovieApi/Models/GetMovieList/FetchMovieListResponse.cs
@@ -1,7 +1,0 @@
-﻿namespace MovieApi.Models.GetMovieList
-{
-    public class FetchMovieListResponse
-    {
-        public List<MovieSummary> Movies { get; set; } = new List<MovieSummary>();
-    }
-}
