--- MovieApi/MovieApi/Models/GetMovieList/GetMovieListResponse.cs
@@ -1,7 +1,0 @@
-﻿namespace MovieApi.Models.GetMovieList
-{
-    public class GetMovieListResponse
-    {
-        public List<MovieSummaryWithProviders> MovieList { get; set; }
-    }
-}
