--- MovieApi/MovieApi/Models/GetMovieList/GetMovieListResponse2.cs
@@ -1,7 +1,0 @@
-﻿namespace MovieApi.Models.GetMovieList
-{
-    public class GetMovieListResponse2
-    {
-        public List<MovieDetail> MovieList { get; set; }
-    }
-}
