--- MovieApi/MovieApi/Models/GetMovieList/MovieSummaryWithProviders.cs
@@ -1,13 +1,0 @@
-﻿namespace MovieApi.Models.GetMovieList
-{
-    public class MovieSummaryWithProviders
-    {
-        public int ID { get; set; }
-        public string Title { get; set; }
-        public string Year { get; set; }
-        public string Type { get; set; }
-        public string Poster { get; set; }
-
-        public List<MovieProviderDetail> ProviderDetails { get; set; }
-    }
-}
