+++ MovieApi/MovieApi/Models/GetMovieListResponse.cs
@@ -1,0 +1,7 @@
+﻿namespace MovieApi.Models
+{
+    public class GetMovieListResponse
+    {
+        public List<MovieDetail> MovieList { get; set; }
+    }
+}
